+++ frontend/src/lib/utils.spec.ts
@@ -2,7 +2,9 @@ import { it, expect } from 'vitest';
 import { createSlug } from "./utils";
 
 it('creates expected slug', () => {
-    expect(createSlug('SomE stRING')).toBe('some-string');
+    expect(createSlug('WeIrD CAsE')).toBe('weird-case');
+    expect(createSlug('several words spaced out')).toBe('several-words-spaced-out');
+    expect(createSlug('symbol (*&$)#)*!@madness')).toBe('symbol-madness');
 });